requirements.txt
# Core
typer>=0.9.0
pydantic>=2.0.0
PyYAML>=6.0.0
pandas>=2.0.0
pyarrow>=14.0.1
httpx>=0.25.0
dependency-injector>=4.41.0
rich>=13.0.0
aiofiles>=23.0.0

# Development
pytest>=7.0.0
pytest-asyncio>=0.21.0